requirements.txt
# Google Agent Development Kit
google-adk>=1.2.0

# LiteLLM for multi-provider LLM support
litellm>=1.40.0

# LangSmith observability
langsmith>=0.1.0

# Environment variable management
python-dotenv>=1.0.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0
